+++ service/bitbucket/annotator.go
@@ -2,17 +2,10 @@ package bitbucket
 
 import (
 	"context"
-	"crypto/sha256"
-	"fmt"
-	"net/http"
 	"path/filepath"
-	"strings"
 	"sync"
 
-	bbapi "github.com/reviewdog/go-bitbucket"
 	"github.com/reviewdog/reviewdog"
-	"github.com/reviewdog/reviewdog/proto/rdf"
-	"google.golang.org/protobuf/proto"
 )
 
 var _ reviewdog.CommentService = &ReportAnnotator{}
@@ -27,38 +20,36 @@ const (
 
 // ReportAnnotator is a comment service for Bitbucket Code Insights reports.
 //
-// API:
+// Cloud API:
 //  https://developer.atlassian.com/bitbucket/api/2/reference/resource/repositories/%7Bworkspace%7D/%7Brepo_slug%7D/commit/%7Bcommit%7D/reports/%7BreportId%7D/annotations#post
 //  POST /2.0/repositories/{username}/{repo_slug}/commit/{commit}/reports/{reportId}/annotations
+//
+// Server API:
+//  https://docs.atlassian.com/bitbucket-server/rest/5.15.0/bitbucket-code-insights-rest.html#idm288218233536
+//  /rest/insights/1.0/projects/{projectKey}/repos/{repositorySlug}/commits/{commitId}/reports/{key}/annotations
 type ReportAnnotator struct {
-	cli         *bbapi.APIClient
+	cli         APIClient
 	sha         string
 	owner, repo string
 
 	muAnnotations sync.Mutex
 	// store annotations in map per tool name
 	// so we can create report per tool
-	annotations map[string][]bbapi.ReportAnnotation
-	severityMap map[rdf.Severity]string
+	comments map[string][]*reviewdog.Comment
 
 	// wd is working directory relative to root of repository.
 	wd         string
 	duplicates map[string]struct{}
 }
 
-// NewReportAnnotator creates new Bitbucket Report Annotator
-func NewReportAnnotator(cli *bbapi.APIClient, owner, repo, sha string, runners []string) *ReportAnnotator {
+// NewReportAnnotator creates new Bitbucket ReportRequest Annotator
+func NewReportAnnotator(cli APIClient, owner, repo, sha string, runners []string) *ReportAnnotator {
 	r := &ReportAnnotator{
-		cli:         cli,
-		sha:         sha,
-		owner:       owner,
-		repo:        repo,
-		annotations: make(map[string][]bbapi.ReportAnnotation, len(runners)),
-		severityMap: map[rdf.Severity]string{
-			rdf.Severity_INFO:    annotationSeverityLow,
-			rdf.Severity_WARNING: annotationSeverityMedium,
-			rdf.Severity_ERROR:   annotationSeverityHigh,
-		},
+		cli:        cli,
+		sha:        sha,
+		owner:      owner,
+		repo:       repo,
+		comments:   make(map[string][]*reviewdog.Comment, len(runners)),
 		duplicates: map[string]struct{}{},
 	}
 
@@ -68,9 +59,14 @@ func NewReportAnnotator(cli *bbapi.APIClient, owner, repo, sha string, runners [
 		if len(runner) == 0 {
 			continue
 		}
-		r.annotations[runner] = []bbapi.ReportAnnotation{}
+		r.comments[runner] = []*reviewdog.Comment{}
 		// create Pending report for each tool
-		_ = r.createOrUpdateReport(context.Background(), reportID(runner, reporter), reportTitle(runner, reporter), reportResultPending)
+		_ = r.createOrUpdateReport(
+			context.Background(),
+			reportID(runner, reporter),
+			reportTitle(runner, reporter),
+			reportResultPending,
+		)
 	}
 
 	return r
@@ -81,19 +77,19 @@ func NewReportAnnotator(cli *bbapi.APIClient, owner, repo, sha string, runners [
 func (r *ReportAnnotator) Post(_ context.Context, c *reviewdog.Comment) error {
 	c.Result.Diagnostic.GetLocation().Path = filepath.ToSlash(
 		filepath.Join(r.wd, c.Result.Diagnostic.GetLocation().GetPath()))
+
 	r.muAnnotations.Lock()
 	defer r.muAnnotations.Unlock()
-
-	anot := r.annotationFromReviewDogComment(*c)
 
 	// deduplicate event, because some reporters might report
 	// it twice, and bitbucket api will complain on duplicated
 	// external id of annotation
-	if _, ok := r.duplicates[*anot.ExternalId]; !ok {
-		r.annotations[c.ToolName] = append(r.annotations[c.ToolName], anot)
+	commentID := externalIDFromDiagnostic(c.Result.Diagnostic)
+	if _, exist := r.duplicates[commentID]; !exist {
+		r.comments[c.ToolName] = append(r.comments[c.ToolName], c)
+		r.duplicates[commentID] = struct{}{}
 	}
 
-	r.duplicates[*anot.ExternalId] = struct{}{}
 	return nil
 }
 
@@ -103,10 +99,10 @@ func (r *ReportAnnotator) Flush(ctx context.Context) error {
 	defer r.muAnnotations.Unlock()
 
 	// create/update/annotate report per tool
-	for tool, annotations := range r.annotations {
+	for tool, comments := range r.comments {
 		reportID := reportID(tool, reporter)
 		title := reportTitle(tool, reporter)
-		if len(annotations) == 0 {
+		if len(comments) == 0 {
 			// if no annotation, create Passed report
 			if err := r.createOrUpdateReport(ctx, reportID, title, reportResultPassed); err != nil {
 				return err
@@ -120,21 +116,25 @@ func (r *ReportAnnotator) Flush(ctx context.Context) error {
 			return err
 		}
 
-		// send annotations in batches, because of the api max payload size limit
-		for start, annCount := 0, len(annotations); start < annCount; start += annotationsBatchSize {
+		// send comments in batches, because of the api max payload size limit
+		for start, annCount := 0, len(comments); start < annCount; start += annotationsBatchSize {
 			end := start + annotationsBatchSize
 
 			if end > annCount {
 				end = annCount
 			}
 
-			// add annotations to the report
-			_, resp, err := r.cli.ReportsApi.BulkCreateOrUpdateAnnotations(
-				ctx, r.owner, r.repo, r.sha, reportID,
-			).Body(annotations[start:end]).Execute()
+			req := &AnnotationsRequest{
+				Owner:      r.owner,
+				Repository: r.repo,
+				Commit:     r.sha,
+				ReportID:   reportID,
+				Comments:   comments[start:end],
+			}
 
-			if err := checkAPIError(err, resp, http.StatusOK); err != nil {
-				return fmt.Errorf("bitbucket.BulkCreateOrUpdateAnnotations: %s", err)
+			err := r.cli.CreateOrUpdateAnnotations(ctx, req)
+			if err != nil {
+				return err
 			}
 		}
 	}
@@ -142,74 +142,27 @@ func (r *ReportAnnotator) Flush(ctx context.Context) error {
 	return nil
 }
 
-func (r *ReportAnnotator) annotationFromReviewDogComment(c reviewdog.Comment) bbapi.ReportAnnotation {
-	a := bbapi.NewReportAnnotation()
-
-	// TODO: allow providing different annotation types in future
-	a.SetAnnotationType(annotationTypeCodeSmell)
-	// hash the output of linter and use it as external id
-	a.SetExternalId(externalIDFromDiagnostic(c.Result.Diagnostic))
-	a.SetSummary(c.Result.Diagnostic.GetMessage())
-	a.SetDetails(fmt.Sprintf(`[%s] %s`, c.ToolName, c.Result.Diagnostic.GetMessage()))
-	a.SetLine(c.Result.Diagnostic.GetLocation().GetRange().GetStart().GetLine())
-	a.SetPath(c.Result.Diagnostic.GetLocation().GetPath())
-	if v, ok := r.severityMap[c.Result.Diagnostic.GetSeverity()]; ok {
-		a.SetSeverity(v)
-	}
-	if link := c.Result.Diagnostic.GetCode().GetUrl(); link != "" {
-		a.SetLink(link)
-	}
-
-	return *a
-}
-
 func (r *ReportAnnotator) createOrUpdateReport(ctx context.Context, id, title, reportStatus string) error {
-	var report = bbapi.NewReport()
-	report.SetTitle(title)
-	// TODO: different report types?
-	report.SetReportType(reportTypeBug)
-	report.SetReporter(reporter)
-	report.SetLogoUrl(logoURL)
-	report.SetResult(reportStatus)
+	req := &ReportRequest{
+		ReportID:   id,
+		Owner:      r.owner,
+		Repository: r.repo,
+		Commit:     r.sha,
+		Type:       reportTypeBug,
+		Title:      title,
+		Reporter:   reporter,
+		Result:     reportStatus,
+		LogoURL:    logoURL,
+	}
+
 	switch reportStatus {
 	case reportResultPassed:
-		report.SetDetails("Great news! Reviewdog couldn't spot any issues!")
+		req.Details = "Great news! Reviewdog couldn't spot any issues!"
 	case reportResultPending:
-		report.SetDetails("Please wait for Reviewdog to finish checking your code for issues.")
+		req.Details = "Please wait for Reviewdog to finish checking your code for issues."
 	default:
-		report.SetDetails("Woof-Woof! This report generated for you by reviewdog.")
+		req.Details = "Woof-Woof! This report generated for you by reviewdog."
 	}
 
-	_, resp, err := r.cli.ReportsApi.CreateOrUpdateReport(ctx, r.owner, r.repo, r.sha, id).Body(*report).Execute()
-
-	if err := checkAPIError(err, resp, http.StatusOK); err != nil {
-		return fmt.Errorf("bitbucket.CreateOrUpdateReport: %s", err)
-	}
-
-	return nil
-}
-
-func hash(b []byte) string {
-	h := sha256.New()
-	_, _ = h.Write(b)
-	return fmt.Sprintf("%x", h.Sum(nil))
-}
-
-// Note that it might be good to create external ID from the diagnostic
-// content along with *original line* (by using git blame for example) to
-// generate unique ID, but it hashes the Diagnostic message for simplicity.
-func externalIDFromDiagnostic(d *rdf.Diagnostic) string {
-	b, err := proto.Marshal(d)
-	if err != nil {
-		b = []byte(d.OriginalOutput)
-	}
-	return hash(b)
-}
-
-func reportID(ids ...string) string {
-	return strings.ReplaceAll(strings.ToLower(strings.Join(ids, "-")), " ", "_")
-}
-
-func reportTitle(tool, reporter string) string {
-	return fmt.Sprintf("[%s] %s report", tool, reporter)
+	return r.cli.CreateOrUpdateReport(ctx, req)
 }