+++ service/bitbucket/annotator_test.go
@@ -2,177 +2,207 @@ package bitbucket
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"reflect"
-	"strings"
 	"testing"
-	"time"
 
-	bbapi "github.com/reviewdog/go-bitbucket"
-	"github.com/reviewdog/reviewdog"
 	"github.com/reviewdog/reviewdog/filter"
 	"github.com/reviewdog/reviewdog/proto/rdf"
+
+	"github.com/reviewdog/reviewdog"
+
+	"github.com/stretchr/testify/suite"
 )
 
-func TestAnnotator(t *testing.T) {
-	var testcases = []struct {
-		name                string
-		runnersList         []string
-		comments            []*reviewdog.Comment
-		expectedAnnotations map[string]int
-		expectedReportCalls map[string][]string
-	}{
-		{
-			name:                "Empty runners list, no comments",
-			expectedAnnotations: make(map[string]int),
-			expectedReportCalls: make(map[string][]string),
-		},
-		{
-			name:                "Predefined runners list, and no comments",
-			runnersList:         []string{"runner1", "runner2"},
-			expectedAnnotations: make(map[string]int),
-			expectedReportCalls: map[string][]string{
-				// since we know list of runners, annotator will create one report for each
-				// - on start it will create reports in "pending" state
-				// - in the end, it will update report to the "passed" state, since no annotations were found
-				reportID("runner1", reporter): {reportResultPending, reportResultPassed},
-				reportID("runner2", reporter): {reportResultPending, reportResultPassed},
-			},
-		},
-		{
-			name:        "Predefined runners list, and one comment",
-			runnersList: []string{"runner1", "runner2"},
-			comments: []*reviewdog.Comment{
-				newComment("runner2", "main.go", "test", 1),
-			},
-			expectedAnnotations: map[string]int{
-				reportID("runner2", reporter): 1,
-			},
-			expectedReportCalls: map[string][]string{
-				// runner1 has no annotations, so report will be marked as "passed"
-				reportID("runner1", reporter): {reportResultPending, reportResultPassed},
-				// runner2 has some annotations, so report will be marked as "failed"
-				reportID("runner2", reporter): {reportResultPending, reportResultFailed},
-			},
-		},
-		{
-			name:        "Predefined runners list, and duplicated comment",
-			runnersList: []string{"runner1", "runner2"},
-			comments: []*reviewdog.Comment{
-				newComment("runner2", "main.go", "test", 1),
-				newComment("runner2", "main.go", "test", 1),
-			},
-			expectedAnnotations: map[string]int{
-				reportID("runner2", reporter): 1,
-			},
-			expectedReportCalls: map[string][]string{
-				reportID("runner1", reporter): {reportResultPending, reportResultPassed},
-				reportID("runner2", reporter): {reportResultPending, reportResultFailed},
-			},
-		},
-		{
-			name:        "Predefined runners list, and many comments",
-			runnersList: []string{"runner1", "runner2"},
-			comments: func() (comments []*reviewdog.Comment) {
-				for i := 0; i < 333; i++ {
-					comments = append(comments, newComment("runner1", "main.go", "test", int32(i)))
-				}
-				return
-			}(),
-			expectedAnnotations: map[string]int{
-				reportID("runner1", reporter): 333,
-			},
-			expectedReportCalls: map[string][]string{
-				reportID("runner1", reporter): {reportResultPending, reportResultFailed},
-				reportID("runner2", reporter): {reportResultPending, reportResultPassed},
-			},
-		},
+type AnnotatorTestSuite struct {
+	suite.Suite
+	cli   *MockAPIClient
+	owner string
+	repo  string
+	sha   string
+}
+
+func (s *AnnotatorTestSuite) SetupTest() {
+	s.cli = &MockAPIClient{}
+	s.owner = "test-owner"
+	s.repo = "test-repo"
+	s.sha = "test-commit"
+}
+
+// Empty runners list, no comments
+func (s *AnnotatorTestSuite) TestEmptyRunnersList() {
+	ctx, annotator := s.createAnnotator(nil)
+
+	err := annotator.Flush(ctx)
+	s.Require().NoError(err)
+	s.cli.AssertExpectations(s.T())
+}
+
+// Predefined runners list, and no comments
+func (s *AnnotatorTestSuite) TestNoComments() {
+	runners := []string{"runner1", "runner2"}
+	ctx, annotator := s.createAnnotator(runners)
+
+	for _, runner := range runners {
+		s.assumeReportCreated(ctx, runner, reportResultPassed)
 	}
 
-	username := "test_user"
-	repo := "test_repo"
-	commit := "test_123"
-	urlPrefix := fmt.Sprintf("/repositories/%s/%s/commit/%s/reports/", username, repo, commit)
+	err := annotator.Flush(ctx)
 
-	for _, test := range testcases {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			reportCallsSequence := map[string][]string{}
-			annotations := map[string]int{}
-			ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				if !strings.HasPrefix(r.RequestURI, urlPrefix) {
-					t.Error("Bad request URI", r.RequestURI)
-					w.WriteHeader(http.StatusBadRequest)
-					return
-				}
+	s.Require().NoError(err)
+	s.cli.AssertExpectations(s.T())
+}
 
-				splt := strings.Split(r.RequestURI, "/")
-				if l := len(splt); l < 8 {
-					t.Error("Request URI should have 7 parts at least, but got", l)
-					w.WriteHeader(http.StatusBadRequest)
-					return
-				}
-				reportID := splt[7]
-				isPostAnnotationsCall := len(splt) >= 9 && splt[8] == "annotations"
+// Predefined runners list, and one comment
+func (s *AnnotatorTestSuite) TestOneComment() {
+	runners := []string{"runner1", "runner2"}
+	comments := []*reviewdog.Comment{
+		s.buildComment(runners[1], "main.go", "test", 1),
+	}
 
-				body, err := ioutil.ReadAll(r.Body)
-				if err != nil {
-					t.Error(err)
-					w.WriteHeader(http.StatusInternalServerError)
-					return
-				}
+	ctx, annotator := s.createAnnotator(runners)
+	s.setupExpectedAPICalls(ctx, runners, comments)
 
-				if isPostAnnotationsCall {
-					var req []bbapi.ReportAnnotation
-					if err := json.Unmarshal(body, &req); err != nil {
-						t.Error(err)
-						w.WriteHeader(http.StatusInternalServerError)
-					}
-					// count how many annotations we created
-					annotations[reportID] += len(req)
-				} else {
-					var req bbapi.Report
-					if err := json.Unmarshal(body, &req); err != nil {
-						t.Error(err)
-						w.WriteHeader(http.StatusInternalServerError)
-					}
-					reportCallsSequence[reportID] = append(reportCallsSequence[reportID], *req.Result)
-				}
+	for _, comment := range comments {
+		err := annotator.Post(ctx, comment)
+		s.Require().NoError(err)
+	}
 
-				w.WriteHeader(http.StatusOK)
-			}))
-			defer ts.Close()
+	err := annotator.Flush(ctx)
 
-			client := NewAPIClientWithConfigurations(&http.Client{Timeout: 1 * time.Second}, bbapi.ServerConfiguration{URL: ts.URL})
-			bb := NewReportAnnotator(client, username, repo, commit, test.runnersList)
-			ctx := context.Background()
+	s.Require().NoError(err)
+	s.cli.AssertExpectations(s.T())
+}
 
-			for _, c := range test.comments {
-				if err := bb.Post(ctx, c); err != nil {
-					t.Error(err)
-				}
+// Predefined runners list, and duplicated comment
+func (s *AnnotatorTestSuite) TestDuplicateComments() {
+	runners := []string{"runner1", "runner2"}
+	comments := []*reviewdog.Comment{
+		s.buildComment(runners[1], "main.go", "test", 1),
+		s.buildComment(runners[1], "main.go", "test", 1),
+	}
+
+	ctx, annotator := s.createAnnotator(runners)
+	s.setupExpectedAPICalls(ctx, runners, comments)
+
+	for _, comment := range comments {
+		err := annotator.Post(ctx, comment)
+		s.Require().NoError(err)
+	}
+
+	err := annotator.Flush(ctx)
+
+	s.Require().NoError(err)
+	s.cli.AssertExpectations(s.T())
+}
+
+// Predefined runners list, and duplicated comment
+func (s *AnnotatorTestSuite) TestManyComments() {
+	runners := []string{"runner1", "runner2"}
+
+	comments := make([]*reviewdog.Comment, 333)
+	for idx := 0; idx < 333; idx++ {
+		comments[idx] = s.buildComment(runners[1], "main.go", "test", int32(idx))
+	}
+
+	ctx, annotator := s.createAnnotator(runners)
+	s.setupExpectedAPICalls(ctx, runners, comments)
+
+	for _, comment := range comments {
+		err := annotator.Post(ctx, comment)
+		s.Require().NoError(err)
+	}
+
+	err := annotator.Flush(ctx)
+
+	s.Require().NoError(err)
+	s.cli.AssertExpectations(s.T())
+}
+
+func (s *AnnotatorTestSuite) createAnnotator(runners []string) (context.Context, *ReportAnnotator) {
+	ctx := context.Background()
+
+	for _, runner := range runners {
+		s.assumeReportCreated(ctx, runner, reportResultPending)
+	}
+
+	annotator := NewReportAnnotator(s.cli, s.owner, s.repo, s.sha, runners)
+
+	return ctx, annotator
+}
+
+func (s *AnnotatorTestSuite) setupExpectedAPICalls(
+	ctx context.Context,
+	runners []string,
+	comments []*reviewdog.Comment,
+) {
+	commentsMap := s.splitComments(comments)
+
+	for _, runner := range runners {
+		expResult := reportResultPassed
+		if len(commentsMap[runner]) > 0 {
+			expResult = reportResultFailed
+		}
+		s.assumeReportCreated(ctx, runner, expResult)
+
+		for start, annCount := 0, len(commentsMap[runner]); start < annCount; start += annotationsBatchSize {
+			end := start + annotationsBatchSize
+
+			if end > annCount {
+				end = annCount
 			}
-			if err := bb.Flush(ctx); err != nil {
-				t.Error(err)
-			}
 
-			// assert resulted calls
-			if !reflect.DeepEqual(test.expectedAnnotations, annotations) {
-				t.Errorf("Expected annotations\n%#v\nbut got\n%#v", test.expectedAnnotations, annotations)
-			}
-
-			if !reflect.DeepEqual(test.expectedReportCalls, reportCallsSequence) {
-				t.Errorf("Expected report calls\n%#v\nbut got\n%#v", test.expectedReportCalls, reportCallsSequence)
-			}
-		})
+			s.assumeAnnotationsCreated(ctx, runner, commentsMap[runner][start:end])
+		}
 	}
 }
 
-func newComment(toolName, file, message string, line int32) *reviewdog.Comment {
+func (s *AnnotatorTestSuite) assumeReportCreated(ctx context.Context, runner string, status string) {
+	s.cli.On("CreateOrUpdateReport", ctx, s.buildReportReq(runner, status)).Return(nil).Once()
+}
+
+func (s *AnnotatorTestSuite) assumeAnnotationsCreated(
+	ctx context.Context,
+	runner string,
+	comments []*reviewdog.Comment,
+) {
+	s.cli.On("CreateOrUpdateAnnotations", ctx, s.buildAnnotationsRequest(runner, comments)).Return(nil).Once()
+}
+
+func (s *AnnotatorTestSuite) buildReportReq(runner string, result string) *ReportRequest {
+	report := &ReportRequest{
+		ReportID:   reportID(runner, reporter),
+		Owner:      s.owner,
+		Repository: s.repo,
+		Commit:     s.sha,
+		Type:       reportTypeBug,
+		Title:      reportTitle(runner, reporter),
+		Reporter:   reporter,
+		Result:     result,
+		LogoURL:    logoURL,
+	}
+
+	switch result {
+	case reportResultPassed:
+		report.Details = "Great news! Reviewdog couldn't spot any issues!"
+	case reportResultPending:
+		report.Details = "Please wait for Reviewdog to finish checking your code for issues."
+	default:
+		report.Details = "Woof-Woof! This report generated for you by reviewdog."
+	}
+
+	return report
+}
+
+func (s *AnnotatorTestSuite) buildAnnotationsRequest(runner string, comments []*reviewdog.Comment) *AnnotationsRequest {
+	return &AnnotationsRequest{
+		Owner:      s.owner,
+		Repository: s.repo,
+		Commit:     s.sha,
+		ReportID:   reportID(runner, reporter),
+		Comments:   comments,
+	}
+}
+
+func (s *AnnotatorTestSuite) buildComment(toolName, file, message string, line int32) *reviewdog.Comment {
 	return &reviewdog.Comment{
 		ToolName: toolName,
 		Result: &filter.FilteredDiagnostic{
@@ -187,4 +217,23 @@ func newComment(toolName, file, message string, line int32) *reviewdog.Comment {
 			},
 		},
 	}
+}
+
+func (s *AnnotatorTestSuite) splitComments(comments []*reviewdog.Comment) map[string][]*reviewdog.Comment {
+	commentsMap := make(map[string][]*reviewdog.Comment)
+	duplicates := make(map[string]struct{})
+
+	for _, comment := range comments {
+		externalID := externalIDFromDiagnostic(comment.Result.Diagnostic)
+		if _, exist := duplicates[externalID]; !exist {
+			commentsMap[comment.ToolName] = append(commentsMap[comment.ToolName], comment)
+			duplicates[externalID] = struct{}{}
+		}
+	}
+
+	return commentsMap
+}
+
+func TestAnnotatorTestSuite(t *testing.T) {
+	suite.Run(t, &AnnotatorTestSuite{})
 }